move source files to src/
