+++ src/delegateAll/index.js
